requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
requests>=2.32.0
python-dotenv>=1.0.1
python-multipart>=0.0.18
